Fix some errors, pytest runs!
